Fix some incorrect implicit float-to-int conversions.

+++ src/camera/basic_first_person_camera.cpp
@@ -10,7 +10,7 @@ hcam::BasicFirstPersonCamera::BasicFirstPersonCamera() :
         if (m_state.aspect_ratio != new_aspect_ratio)
             set_aspect_ratio(new_aspect_ratio);
     })),
-    m_clamp_up({true, 60.0f / 360.0f * 2.0f * M_PI})
+    m_clamp_up({true, 60.0f / 360.0f * 2.0f * static_cast<f32>(M_PI)})
 { /* Empty. */ }
 
 void hcam::BasicFirstPersonCamera::update() {
